requirements.txt
Flask==1.1.1
Flask-WTF==0.14.2 
Flask-RESTful==0.3.7
Flask-Login==0.4.1
Flask-JWT-Extended==3.24.1
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.2
Flask-Uploads==0.2.1
psycopg2==2.8.4
gunicorn==19.9.0 
requests==2.22.0
dateparser==0.7.2
pytz==2019.3
icalendar==4.0.3
hashids==1.2.0
python-slugify==4.0.0
python-dateutil==2.8.1
boto3==1.10.28
markdown2==2.3.8
mixpanel==4.5.0